⭐created a new app called `accounts`
