+++ rb_backend/accounts/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+]
